LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-cancel-events-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: action
Test cancel event is fired when the dialog is closed by user interaction

Hello World


FAIL Test cancel event is fired when the dialog is closed by user interaction ReferenceError: Can't find variable: action
